fix (i hope) the tmpwm race condition and the bug that sometimes keys don't work after tmpwm
the race condition is fixed by spinning in a loop with a special error
handler until it is confirmed that the root event selections were
successfully.

Existing windows don't get their top level keys grabbed, so after the
windows are synced up, grab the top level keys on all windows.