cmd_info: also display a window not in the current group
Without this change, cmd_info displays "No Window" if called
with the current window not in the current group (for example
directly after a gselect).

This change makes it also look in other groups. In that case
it might show a number not the one to switch back to it without
switching the group first, but I guess that is less confusing
than just claiming there is no window.