Allow exact matching with gselect.
When having two groups where the second one's name partially matches the name
of the first one, it is hard to gselect the second one. Similarly it is not
possible to select groups with fully numerical names.

e.g.

If having the group list:

 0-default
 1*other
 2-de
 3-0

The user expactation when typing ":gselect de" or ":gselect 0" would likely be
to select group number 2 & group number 3 respectively.

This commit modifies the behaviour of gselect, to primarily select the group
with an exactly matching name if possible, before trying numerical or partial
matches.