Update screen-aware focus* commands after changes to support XRandR
This is an update to Pedro Silva's patch, so that it applies against the
xrandr branch.  That patch introduced four new functions, frame_*_abs,
which report a frame's left, right, top, and bottom coordinates offset by
it's screen arrangement.  This allows the use of the focus* commands as
additional screen navigation tools, and has the added benefit of making the
find_frame_* functions clearer.