Fix: call using_history after initializing history entries
In its current state, history expansion requiring any sort of string
lookup is not working, always returing "event not found". This is due
to `history_offset` being set improperly by an early call to
`using_history`.

A call to `using_history` needs to follow, not precede, the loop
performing initialization of history entries found in
`history_expand_line`. This allows for `history_offset` to reflect the
now non-zero length of the history, restoring the expected behaviour
of history expansion.

Tested-by: Martin Hertz <mvhertz@gmail.com>
Signed-off-by: Antoine Busque <antoinebusque@gmail.com>