Fix: sync wins correctly after tmpwm for multiple screens
The current implementation of `sync_wins`, which gets called after
control is returned to ratpoison following a call to `tmpwm`, gets
executed once per screen.

This is not only unnecesary, but also causes ratpoison to break after
returning from `tmpwm`. This is due to two things: first, we can scan
for the windows in only one pass, because the call to `XQueryTree`
returns all windows below the root window, which is shared across the
screens. Second, the current implementation only checks whether a
window is a special ratpoison window (i.e. the key window, bar window,
etc.) for the current screen. This results in some windows being
mapped although they shouldn't be, which is what causes the break
preventing ratpoison from receiving commands correctly,

This patch fixes these issues by only running `sync_wins` once instead
of on all screens, and by using the existing `is_rp_window` utility
function to prevent a special ratpoison window from being mapped,
regardless of the screen it belongs to.

Signed-off-by: Antoine Busque <antoinebusque@gmail.com>