add onlyborder variable
When onlyborder is set to 0 (default 1), borders are not displayed if
there is only a single frame on a given screen.

This is useful if you use (thick) coloured borders to show which is the
currently focused window when using splits, but also don't want to have
borders taking up space unnecessarily when running a fullscreen window.

Signed-off-by: Vasil Zlatanov <v@skozl.com>