Add support for virtual screens
Frame and groups are now held by an rp_vscreen object, which is held
in the vscreens list in each rp_screen.  Bar/help/etc. windows stay
associated with the rp_screen since they are present on every
virtual screen.

A vscreens variable determines how many virtual screens to create,
which defaults to 5.

The vselect command switches to a given virtual screen.

The vmove command moves the current window to the given virtual
screen.