Add ignoreresizehints setting
When enabled, PResizeInc hints will be ignored on windows, forcing
them to be the exact size of their containing frame.

This is useful for terminals which specify a PResizeInc hint to
request that their size be a multiple of the terminal's font size.
While ignoring this hint will cause the terminal to have padding
around the bottom/right sides, it may be better looking than having
the window the "correct" size and having that empty space be in the
frame.