Add support for full-screen windows
When a _NET_WM_STATE message comes in and the window requests
_NET_WM_STATE_FULLSCREEN, hide the bar and make the window take over
the whole screen out of a frame.

When a focus or vscreen change happens, un-fullscreen.

Works with Chrome and VLC, though neither seems to respond to the
message forcing them out of full-screen mode, so they just end up
resizing into a frame but not realizing they should go back to a
normal window.  I need to confirm whether that is the expected
behavior.

There should be a setting to control whether full-screen windows are
allowed, or maybe piggy back on rudeness.