Add support for drawing arbitrary text in the sticky bar
Create a FIFO at ~/.ratpoison_bar and listen for any input on it.
When a newline-terminated line has been received, display it on the
right side of the sticky bar.

This provides a more integrated bar than using something like dzen2.

Being able to parse dzen's control functions like "^fg(#abcdef)" for
per-segment colors would be a nice bonus.