Overhaul IPC mechanism to use a Unix socket for increased security
The Atom-based IPC inherited from Ratpoison is overly complex and
insecure.  An unprivileged application that only has an X11
connection (say a sandboxed Firefox child process) could set the
RP_COMMAND_REQUEST atom property on its window with a value of
"0exec something" to cause us to read it and execute that shell
command.

Switching to a Unix socket in our ~/.config/sdorfehs directory
ensures the requesting process is running as our uid, has the
ability to make socket connections, and can write to that path.