cmd_execf: Make this actually work (again?)
When spawn()ing a command, it is executed as an argument to /bin/sh
which means the pid that fork() returned in spawn() is the pid of
sh, not the program sh ultimately executes.  Work around this by
passing the command as "exec command" which tells sh to exec it
directly instead of forking it.

I'm not sure how this ever worked since it seems the same in
ratpoison.

Also remove the raw argument to spawn since nothing uses it.