get_more_input: Work around a weird issue by sleeping
When doing a command like ":vselect" with no args, the input window
will be hidden, the sticky bar text will be redrawn, then
get_more_input is called again to prompt the user for a vscreen so
the bar window is hidden and the input window is shown again but
that second input window doesn't show up until the user types a
character.

I don't know why this sequence triggers it, but sleeping for a very
brief moment allows the window to show.  I have no idea why.