bar: Allow multiple buttons/commands in a single ^ca()
This isn't supported by dzen2 but it seems like a natural extension.

It allows a single string of data like a speaker icon, but multiple
commands depending on the button clicked, such as a volume up action
when button 4 is clicked (mouse wheel scrolling up), and a volume
down action when button 5 is clicked (mouse wheel scrolling down).

Defer parsing these until we get an actual click since the parsing
is a bit more complicated now and we're probably going to have many
more updates to the bar than mouse clicks.