screen: Use screen_primary() to find the bar, not xrandr_is_primary()
xrandr might not be available or we might not have a primary screen.
screen_primary() factors in xrandr_is_primary() in its calculation
but has better logic for determining an actual primary screen.

Fixes the issue of the bar drawing over the only frame which can be
seen in Xnest.