config: Define VERSION from Makefile, possibly using git commit id
This 20190822 version hasn't been accurate since forking from
ratpoison.

If there is a .git directory, assume we're building from git and
pass "git-XXXXXXX" as the version to help with bug reports, but to
avoid an unnecessary build dependency on git, if there is no .git,
just pass the hard-coded VERSION which will get bumped on each
release.