window: when adding to windowlist, use current vscreen if 'winvcur' set
Add boolean rcvar 'winvcur' (default 0).

In the specific case of a new window where the parent was originally
started in one vscreen, but spawns another window later when a different
vscreen is active, 'winvcur' will cause the new window to be added to
whatever the current vscreen at the time, rather than the ostensible
parent vscreen.

Default behavior (use parent's initial vscreen) is unchanged.

The original impetus for this patch was to fix use of 'urxvtc' with
'urxvtd', coming from old ratpoison "groups" and migrating to vscreens.
'urxvt' standalone already worked as expected: starting in the current
vscreen, since these are completely new processes.  However 'urxvtc' are
child windows all of the same urxvtd process.