window: if msgwait 0 persist and update bar lists on vselect
As code is now, when changing vscreens, the current window is
un-fullscreened (if needed) and bar is hidden.  With this patch, we
don't hide the bar if msgwait is 0.  This leaves the vscreen or windows
list up if it were up already, even when changing vscreens.  It gets
updated -- ie redrawn -- to the new correct windows or vscreen names and
upon switch.  This lets the user keep the list of windows up and move
between either windows or vscreens, while retaining correctly updated
winlist contents.

This seems to retain functionality of fullscreen windows also.  However,
it was only tested with Chrome's fullscreen, which for me worked even
before the EWMH patches.