events: client_msg: support apps that raise via _NET_ACTIVE_WINDOW
some applications like Chrome, in the tab selection window, if they
think the WM is EWMH-compliant[*], will send a _NET_ACTIVE_WINDOW
ClientMessage with the target window ID of the one it wants raised.

in eb58c16a support was added for _NET_ACTIVE_WINDOW but this is only a
property maintained for others to read.  this patch will also support
receiving it as a ClientMessage and switching windows accordingly.

[*] works in Ratpoison without this patch, so the presumption is
_NET_SUPPORTING_WM_CHECK added to sdorfehs in 95f44a4 is used to decide
what method to use for window switch; this is not proven but nonetheless
this patch works to restore functionality