get_child_info: use XResQueryClientIds() if _NET_WM_PID not found
Using server-maintained PID extension is a lot more reliable as it's
directly tracked by the X server and no need for client to set anything.
We still use _NET_WM_PID if it's available to retain old behavior and
allow for setting different PID than actual PID which some programs do.

Without this patch, (1) the program must be a "good" conforming client
and always set its own _NET_WM_PID which is voluntary, and (2) it must
do it before the window manager receives the CreateNotify event and
tries to look up the window.  This race is frequently failing for some
applications like `st`.

This does require libXRes 1.2+ and headers to be present on the system
in order to work (on my system, this means libxres-dev and libxres1
packages).  These should be available on anything relatively recent in
last few years, 1.2 was released 5 years ago:

https://gitlab.freedesktop.org/xorg/lib/libxres/-/tree/libXres-1.2.0