actions: cmd_execw: new exec command waits for forked proc to exit
Normally it is not possible to sequence rcfile commands that run
programs, because 'exec' forks and does not wait for completion before
moving on to the next statement in rcfile.  This patch adds an 'execw'
command that will wait indefinitely for the shell to exit.  User should
use timeout(1) if they don't want to hang their window manager.