rudeness: add a bit to allow off-vscreen window raise, default off
before, off-vscreen raises were unconditionally refused.  This is a
problem when using Chrome's Control-Shift-A to switch to a different
window in another vspace, as the request will always be denied, despite
being user-requested (and with no way to allow it).

With the addition of the new bit, rudeness will be consulted when window
in another vscreen requests raise, and vscreen will be switched to, if
allowed.  It defaults to "off" so that:

    - all old behavior is unchanged,
    - old values for rudeness will have the same behavior they always
      did (to deny off-vscreen raise), and
    - not setting rudeness at all will keep the same value as before

Probably, other use cases would be satisfied by this patch as well...