main: use ordinary char type for send_command() arg, not unsigned
probably the intention here was to support characters with values > 127
(ie iso_8859_1 chars), but there's no reason to use unsigned for this:
in C the string functions work just fine as their default type (char)
and don't require casting to support high-ascii characters.

of course, it still won't work with chars > 255, as before.

this change was tested with "sdorfehs -c $'echo \uab'" which correctly
shows the double-arrow, "left-pointing double angle quotation mark"